feat: centralize cursor rendering logic in CursorRendererHelper
